Refactor LoginNep2
